+++ packages/Annotation/README.md
@@ -1,0 +1,11 @@
+# Annotation Module
+
+## How to Render Custom Annotation
+
+- @todo AnnotationSubscriberInterface
+
+## How to Modify Description
+
+For custom markdown etc. rendering
+
+- @todo Description Event
