+++ packages/Annotation/src/AnnotationSubscriber/LinkAnnotationSubscriber.php
