+++ packages/Annotation/tests/AnnotationSubscriber/SeeAnnotationSubscriberTest.php
@@ -1,13 +1,13 @@
 <?php declare(strict_types=1);
 
-namespace ApiGen\Tests\Annotation\AnnotationSubscriber;
+namespace ApiGen\Annotation\Tests\AnnotationSubscriber;
 
 use ApiGen\Annotation\AnnotationDecorator;
 use ApiGen\Annotation\AnnotationList;
 use ApiGen\Reflection\Contract\Reflection\Class_\ClassMethodReflectionInterface;
 use ApiGen\Reflection\Contract\Reflection\Class_\ClassReflectionInterface;
 use ApiGen\Tests\AbstractParserAwareTestCase;
-use ApiGen\Tests\Annotation\AnnotationSubscriber\SeeAnnotationSubscriberSource\SomeClassWithSeeAnnotations;
+use ApiGen\Annotation\Tests\AnnotationSubscriber\SeeAnnotationSubscriberSource\SomeClassWithSeeAnnotations;
 
 final class SeeAnnotationSubscriberTest extends AbstractParserAwareTestCase
 {
@@ -50,7 +50,7 @@ final class SeeAnnotationSubscriberTest extends AbstractParserAwareTestCase
         $seePropertyAnnotation = $this->methodReflection->getAnnotation(AnnotationList::SEE)[2];
 
         $this->assertSame(
-            '<code><a href="class-ApiGen.Tests.Annotation.AnnotationSubscriber.SeeAnnotationSubscriberSource.PresentReturnedClass.html#$someProperty">PresentReturnedClass::$someProperty</a></code>',
+            '<code><a href="class-ApiGen.Annotation.Tests.AnnotationSubscriber.SeeAnnotationSubscriberSource.PresentReturnedClass.html#$someProperty">PresentReturnedClass::$someProperty</a></code>',
             $this->annotationDecorator->decorate($seePropertyAnnotation, $this->methodReflection)
         );
     }
@@ -70,7 +70,7 @@ final class SeeAnnotationSubscriberTest extends AbstractParserAwareTestCase
         $seeMethodAnnotation = $this->methodReflection->getAnnotation(AnnotationList::SEE)[3];
 
         $this->assertSame(
-            '<code><a href="class-ApiGen.Tests.Annotation.AnnotationSubscriber.SeeAnnotationSubscriberSource.PresentReturnedClass.html#_someMethod">PresentReturnedClass::someMethod()</a></code>',
+            '<code><a href="class-ApiGen.Annotation.Tests.AnnotationSubscriber.SeeAnnotationSubscriberSource.PresentReturnedClass.html#_someMethod">PresentReturnedClass::someMethod()</a></code>',
             $this->annotationDecorator->decorate($seeMethodAnnotation, $this->methodReflection)
         );
     }
@@ -90,7 +90,7 @@ final class SeeAnnotationSubscriberTest extends AbstractParserAwareTestCase
         $seeFunctionAnnotation = $this->methodReflection->getAnnotation(AnnotationList::SEE)[5];
 
         $this->assertSame(
-            '<code><a href="function-ApiGen.Tests.Annotation.AnnotationSubscriber.SeeAnnotationSubscriberSource.someExistingFunction.html">someExistingFunction()</a></code>',
+            '<code><a href="function-ApiGen.Annotation.Tests.AnnotationSubscriber.SeeAnnotationSubscriberSource.someExistingFunction.html">someExistingFunction()</a></code>',
             $this->annotationDecorator->decorate($seeFunctionAnnotation, $this->methodReflection)
         );
     }
