+++ packages/Annotation/tests/AnnotationSubscriber/Source/UsedClass.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace ApiGen\Annotation\Tests\AnnotationSubscriber\Source;
+
+final class UsedClass
+{
+}
