--- tests/Annotation/AnnotationSubscriber/Source/UsedClass.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace ApiGen\Tests\Annotation\AnnotationSubscriber\Source;
-
-final class UsedClass
-{
-}
